Restructure `electrodes` data structure
The current structure of `electrodes` is three Arrays: `initPositions`,
`deltas`, and `ids`. The idea is that each element at each index of each
array corresponds with a certain electrode. However, due to errors
within the code, it is very easy for the three Arrays to become
'disconnected'. This is because the code would try to match an
electrode's id from `ids` to its corresponding `initPositions` and
`deltas` by using the index of the Arrays. However, when electrodes get
deleted, the corresponding elements in each Array would get deleted.
However, this meant that now the indices in `initPositions` and `deltas`
no longer correspond with the entries in `ids`.

While working through this problem, it became evident to me that this
data structure just doesn't really make any sense. Instead, `electrodes`
should just be one Array, and instead each element would contain the
electrode's corresponding `initPositions`, `deltas`, and `ids`. That
way, it is impossible for each electrode's data to become disconnected,
because either the electrode exists and all three data members exist, or
the electrode doesn't exist and none of its data members exist.

This commit restructures `electrodes` to match the above structure, thus
eliminating the issue in #238 as well as potentially #241.

NOTE: From my understanding, the `ids` do not try to minimize
themselves. Previously, the `ids` array did not do this, but because the
indices of Arrays always range from [0, length-1], the effective `ids`
of `initPositions` and `deltas` did minimize. These changes restore the
intended design of the ids. Now, when new electrodes are added, the
`ids` are assigned from the highest id + 1. This means that when
electrodes are deleted, their ids are not reused. This means that the
ids can become very larger even if the number of electrodes is low.
However, I don't believe that this is an actual issue.